Remove duplicate ValueListBuilder implementation
